+++ MovieRental.DAL/Services/FilmService.cs
@@ -19,9 +19,9 @@ namespace MovieRental.DAL.Services
                 (int)reader["ReleaseYear"], 
                 (int)reader["LanguageId"], 
                 (int)reader["RentalDuration"], 
-                (double)reader["RentalPrice"], //Might have a problem going from decimal (5,2) to double
+                (decimal)reader["RentalPrice"], //Might have a problem going from decimal (5,2) to double
                 (int)reader["Lenght"], 
-                (double)reader["ReplacementCost"], 
+                (decimal)reader["ReplacementCost"], 
                 (int)reader["RatingId"]
             );
         }
